최종 트래킹 화면 GUI디자인 업데이트, 설정 부분 오류 해결 예정
